feat: centralize all user routes to single router and service layer
