+++ backend/main.py
@@ -4,7 +4,7 @@ from fastapi.middleware.cors import CORSMiddleware
 from sqlalchemy.orm import Session
 
 from database import get_db
-from routers import admin, admin_user, user
+from routers import admin, user
 from scripts import (
     _run_end_of_day_cron_job,
     get_scheduler,
@@ -65,4 +65,3 @@ def eod_cron_job(db: Session = Depends(get_db)):
 
 app.include_router(admin.router)
 app.include_router(user.router)
-app.include_router(admin_user.router)
